feat(core,bench): f32 render geometry + baseline series fork — slice 3 of f32-hot-columns
Render batch NumColumn/NumOpt store f32; both fill paths (compiled
batch/row and interpreted row walk) round geometry and num extras
through f32 at entry, so batch-vs-row and compiled-vs-interp compares
stay bit-exact while RenderRow keeps its f64-typed host API carrying
f32-exact values. Mesh-pack casts collapse where upstream is already
f32-exact; fits_f32 guards and FRAME_ABI_VERSION unchanged. Bench
series forks maku-v1-f64 -> maku-v1-f32 across native/browser/failure
producers and matrix json; results/maku-v1-f64 evidence untouched.

Gates first-hand: 350 core tests plain + MAKU_LOWER_ORACLE=1, 6
release ignored oracle suites, workspace build, native bench smoke
(4 tiers, 39 schema documents OK).

+++ crates/core/src/interp/engine.rs
@@ -467,12 +467,12 @@ impl RenderRowFields {
         };
         let data = match &*shape {
             "point" | "dot" => RenderData::Point {
-                x: self.x.map(|v| v.num()).transpose()?.unwrap_or(0.0),
-                y: self.y.map(|v| v.num()).transpose()?.unwrap_or(0.0),
-                theta: self.theta.map(|v| v.num()).transpose()?.unwrap_or(0.0),
-                scale: self.scale.map(|v| v.num()).transpose()?.unwrap_or(1.0),
-                alpha: self.alpha.map(|v| v.num()).transpose()?.unwrap_or(1.0),
-                hue: self.hue.map(|v| v.num()).transpose()?.unwrap_or(0.0),
+                x: self.x.map(|v| v.num()).transpose()?.unwrap_or(0.0) as f32 as f64,
+                y: self.y.map(|v| v.num()).transpose()?.unwrap_or(0.0) as f32 as f64,
+                theta: self.theta.map(|v| v.num()).transpose()?.unwrap_or(0.0) as f32 as f64,
+                scale: self.scale.map(|v| v.num()).transpose()?.unwrap_or(1.0) as f32 as f64,
+                alpha: self.alpha.map(|v| v.num()).transpose()?.unwrap_or(1.0) as f32 as f64,
+                hue: self.hue.map(|v| v.num()).transpose()?.unwrap_or(0.0) as f32 as f64,
             },
             "polyline" => {
                 let points = match &shape_value {
@@ -521,7 +521,7 @@ impl RenderRowFields {
             match v {
                 Val::Num(n) => {
                     check(world, &key, RenderFieldKind::Num)?;
-                    row.nums.push((key, n));
+                    row.nums.push((key, n as f32 as f64));
                 }
                 Val::Kw(sym) => {
                     check(world, &key, RenderFieldKind::Sym)?;
